Added getBerryList Task
