+++ TestWebAPI/Models/Berry.cs
@@ -8,5 +8,14 @@ namespace TestWebAPI.Models
     public class Berry
     {
         public String name { get; set; }
+        public string url { get; set; }
+    }
+
+    public class BerryRoot
+    {
+        public int count { get; set; }
+        public string next { get; set; }
+        public object previous { get; set; }
+        public List<Berry> results { get; set; }
     }
 }
